flask/requirements.txt
Flask==2.0.2
Flask_RESTful==0.3.9
Flask_SQLAlchemy==2.5.1
requests==2.28.1
uwsgi==2.0.21